Create fibbonacci.py
This Python program is designed to generate and print the Fibonacci series up to a given number, as well as find the sum of the terms in the series up to that number.his Python program is designed to generate and print the Fibonacci series up to a given number, as well as find the sum of the terms in the series up to that number. In this program, we first take input from the user to determine up to which number we want to print the Fibonacci series. We then initialize the first two terms of the series to 0 and 1 respectively.
Using a while loop, we iterate through the series until we reach the given number. Within the loop, we print out each term of the series and add it to the fib_sum variable. We then update the values of a and b to continue generating the next terms of the series.